LayoutTests/inspector/timeline/timeline-enum-stability.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script>

function test()
{
    InspectorTest.addObject(WebInspector.TimelineModel.RecordType);
    InspectorTest.completeTest();
}

</script>
</head>

<body onload="runTest()">
<p> 
Check to see that TimelineAgent record types are stable.
</p>

<p>
Applications outside of WebKit depend on the stability of the mapping of 
these types to these specific values.
</p>

</body>
</html>
